move emoji insertion to content script
