+++ vite.config.ts
@@ -22,7 +22,7 @@ const ignoreHMR = [
 	"SettingsModule.vue",
 ];
 
-const alwaysHot = ["src/background/background.ts"];
+const alwaysHot = ["src/background/background.ts", "src/content/content.ts", "src/content/emoji.ts"];
 
 // https://vitejs.dev/config/
 export default defineConfig(() => {
